pyproject.toml
[tool.poetry]
name = "lobbybee-backend"
version = "0.1.0"
description = ""
authors = ["darkwebplayer <amrithanadhm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
djangorestframework = "^3.16.0"
django-environ = "^0.12.0"
django-filter = "^25.1"
markdown = "^3.8.2"
django-cors-headers = "^4.7.0"
boto3 = "^1.40.2"
django-storages = "^1.14.6"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.5.1"
celery = "^5.4.0"
redis = "^5.2.0"
django-celery-beat = "^2.7.0"
gunicorn = "^23.0.0"
psycopg = {extras = ["binary"], version = "^3.2.9"}
pyaadhaar = "^2.0.2"
pillow = "^11.3.0"
pyzbar = "^0.1.9"
requests = "^2.32.5"
channels = "^4.3.1"
channels-redis = "^4.3.0"
daphne = "^4.2.1"


[tool.poetry.group.dev.dependencies]
awscli = "^1.42.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
